Generic shares_base module and specific s3_datasets_shares module - part 2 (db objects to shares_base) (#1294)
### Feature or Bugfix
- Refactoring

### Detail
As explained in the design for #1123 and #1283 we are trying to
implement generic `datasets_base` and `shares_base` modules that can be
used by any type of datasets and by any type of shareable object in a
generic way.

In this PR:
- Move shares state machines into its own file and move them to the
`shares_base` module.
- Create new ShareObjectRepository and copy some generic methods to it
- Move db share objects to shares_base:
- ShareObject has a field called datasetUri. For the scope of S3 and
Redshift datasets we can leave it as dataseturi, but if we want to
implement other kinds of sharing we should rethink it: we need to store
the "approvers" of the share in some way. For the moment I am not going
to go down that path until shares and s3 are uncoupled, then we can see
how we would implement a complete generic sharing.
- ShareObjectItem includes 3 fields related to the glue tables... we
need to get rid of them in the backend code and look up the info with
the itemType and itemUri. Left for part3 to keep the PR clean.


### Relates
- #1283 
- #1123 
- #955 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.