Fix s3_datasets and s3_datasets_shares tests (#1325)
### Feature or Bugfix
- Bugfix

### Detail
With the refactoring of the datasets and the datasets_sharing tests were
not being executed. In this PR the corresponding tests packages are
renamed and the tests inside are fixed with the latest changes in main.

There are still things to improve, for example we can simplify the
conftests of the s3_datasets and the s3_datasets_shars modules; but I am
going to leave it for a next PR

### Relates
- #1123 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.