Add search (Autocomplete) in environment/group dropdown in Share request modal (#1335)
### Feature or Bugfix
- Feature

### Detail
In the following view, instead of using a fixed list in the environment,
team and consumption roles dropdowns; this PR introduces search
capabilities as requested in #1012. There are still many other dropdowns
to change, where we will be able to extract a common frontend component.
This one however, is a bit particular so I implemented it separately.


![image](https://github.com/data-dot-all/dataall/assets/71252798/971b8c95-1211-4aa1-a714-0b1548f2ccd9)


### Relates
- #1012 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.