Add search (Autocomplete) in group dropdown for useGroups (#1356)
### Feature or Bugfix
- Feature

### Detail
Autocomplete for environments and teams in the following frontend views
as requested in 1012. In this PR the changes apply to those views that
use `useGroups` to define the `groupOptions`
- [x] WorksheetCreateForm.js
- [x] EnvironmentCreateForm.js
- [x] OrganizationCreateForm.js
- [x] GlossaryCreateForm.js

❗ In the EnvironmentCreateForm we might need to filter out the list of
groups provided. At the moment users can select any group that they
belong to. With the changes done by @SofiaSazonova in #1306 I think we
need to review if that is still appropriate. Do you think we should open
a new PR for this change?

### Relates
- #1012 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.