Add search (Autocomplete) in environment/group dropdown in create/import forms (#1347)
### Feature or Bugfix
- Feature

### Detail
Autocomplete for environments and teams in the following frontend views
as requested in 1012.

This PR introduces a 2 new Shared frontend components:
EnvironmentTeamDropdown and EnvironmentTeamDatasetsDropdown. They are
generic formik fields that can be reused in Formik forms in several
modules. Some creation forms include custom parameters that are fetched
alongside the environments and teams (e.g. Notebooks sets VPCs). For
those cases, we implement the Autocomplete logic in the view itself

- [X] DashboardImportForm.js - using EnvironmentTeamDropdown
- [X] MLStudioCreateForm.js - using EnvironmentTeamDropdown
- [X] NotebookCreateForm.js - custom
- [X] OmicsRunCreateForm.js - using EnvironmentTeamDatasetsDropdown
- [X] DatasetCreateForm.js - custom
- [X] DatasetImportForm.js - custom

Pipelines in a separate PR

In all forms we use `SamlAdminGroupName` instead of `SamlGroupName` as a
standard that allows us to use Shared components

### Relates
- https://github.com/data-dot-all/dataall/pull/1335
- #1012 


### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.